Don't truncate the machineId
An MD5 checksum is 128 bits.  When formatted in hex, that is 32
characters.  But the way we use the MD5 checksum in machineId also
requires space for a terminating NUL character.  Add it.

Luckily, the way we compute this, this wasn't a buffer overflow.  We
just truncated the MD5 checksum after the first 31 characters.

(Actually, since the machineId is computed from chip ID which is only
32 bits long, only 2**32 different values are possible.  It might make
sense to truncate this to 8 characters, or maybe 10 or 12 or even 16
to allow for the birthday paradox; 31 or 32 characters is overkill.
At home, where I only have 10 thermometers, I use 6 characters since
that is more convenient to type.  I could probably get by with just 2
or 4.)