Add support for writing to the display via MQTT
If you publish a string to the "cmnd/<machineId>/line1" topic it will
be displayed on the first line of the display.  You can also write to
the second and third line by using topics that end with line2 and
line3.  To revert to the default display, just publish an empty
string.