Add OTA upgrade support
To upgrade your Anavi Thermometer, place the image on a web server
somewhere.  (It is recommended to use signed images.)  Then publish a
message to the "cmnd/<machineId>/update" topic with a JSON message on
this form:

    {"file": "/anavi.bin", "server": "myserver.example.com"}

The Anavi Thermometer will download the image and reboot to the new
version.

Note: you sometimes have to power cycle the Anavi Thermometer after
the upgrade to get any sensor readings, at least from the DHT22
sensor.