Enable factory reset during boot
The factory reset button only works once the Anavi Thermometer has
successfully connected to a network.  This is problematic if you move
the thermometer to another location.  While the Thermometer will fall
back to the normal network setup if it cannot connect, I have found
that the setup works better when the thermometer has been reset.

This commit blinks rapidly with the status led for 2 seconds during
setup.  If you press the reset button during this time, the normal
factory reset sequence will be started: keep holding the button for 3
seconds while it blinks slowly, and it will apply the factory
settings.