Refactoring: store display sensor values in global variables
Introduce sensor_line1, sensor_line2 and sensor_line3, and use them
instead of the air, hum and water local variables of loop().  The
"water" variable was already named a bit funny, as it sometimes
contained the RSSI value.

Similar to mqtt_lineX, these variable represents the display contents
that the local sensors want to show on the display.

The purpose of this change is to enable the display to be more
responsive to updates via MQTT -- see the next commit.