Redraw the display immediately when it is updated via MQTT
Previously, the display would only be redrawn roughly once every ten
seconds, when the sensors were read.

For this to work, we need to retain the sensor values, so that we can
restore the display when somebody publishes an empty string via MQTT.
Conveniently enough, the previous commit did just that.