Prevent buffer overruns in json.printTo
When json.printTo is invoked with a char* and an int, the int should
describe how big the buffer is.  The ArduinoJson library won't write
more than fits in the buffer.

The json.measureLength() method can be used to calculate how large the
buffer needs to be, and you can then allocate a suitably large buffer.
But in our case, a buffer of 100 bytes is surely large enough, so
there is no need for dynamic allocation.

By just passing a char array to printTo, it will automatically use the
size of the buffer to avoid buffer overruns.  (Alternatively, we could
have used sizeof(payload) as the second argument of json.printTo.)