Support ESP12_BLUE_LED_ALWAYS_ON
If you are annoyed by the flashing blue LED on the ANAVI Thermometer,
you can now define ESP12_BLUE_LED_ALWAYS_ON to enable a hack that will
instead make it be almost always on.  It will still flicker while a
temperature measurement is made, but so quickly that the human eye
can't detect it.

Also documented why the blue LED blinks in the first place.