Allow the DS18B20 port to be used as a button
GPIO12 is normally as a 1-Wire bus by the ANAVI Thermometer.  The
stock firmware supports connecting a single DS18B20 temperature
sensor, but in principle you could connect more 1-Wire sensors.  That
would require some changes to the firmware, though.

This commits add support for using GPIO12 for a totally different
purpose instead: as an input for a button.  The ANAVI Thermometer will
publish an "ON" MQTT message when the button is pressed, and an "OFF"
message when it is released.  (See the comments in the code for
details.)  These messages can be used to trigger a scene or automation
in Home Assistant.

To use this, connect a normally closed switch in series with a 470 ohm
resistor between the GND and DATA pins of the DS18B20.  The code
autodetects that the input is held low when the ANAVI Thermometer
boots, disables the 1-Wire functionality, and starts polling the
switch instead.