Configure altitude and/or sea level pressure via MQTT for BMP180
The BMP180 pressure sensor library can calculate the sea-level
pressure if the altitude and air pressure are known.  It can also
calculate the altitude if the sea-level and air pressure are known.

Stop using the SENSORS_PRESSURE_SEALEVELHPA constant.  Instead, listen
to MQTT topics that allow you to specify the altitude and/or sea level
pressure.  If the altitude is given, the calculated sea level pressure
will be published via MQTT.  If the sea level pressure is given, the
calculated altitude will be published via MQTT.  (It is possible, but
perhaps not useful, to specify both.)

The BMPaltitude MQTT topic is now published only if the sea-level
pressure has been configured via MQTT.