Fix HA Discovery: each device needs to be unique
Home Assistant seems to get confused when the temperature sensor and
pressure sensor have exactly the same device information.  By
appending the name suffix to the name, things seem to work better.

I'm not 100% sure this is the correct thing to do, though.  I have not
been able to find a resource that explains the relationship between
devices, entities, the state machine, and MQTT Discovery in a way that
I fully understand.  That said, this seems to work better.