Avoid losing the DHT22 sensor after restarts
The DHT22 sensor often stops working when the ANAVI Thermometer is
restarted.  This commit fixes this in most, but not all, situations.
In my testing, it is now possible to do a MQTT-initiated restart or
factory reset without losing the sensor.  However, a power cycle is
unfortunately still required after an OTA upgrade.