Don't tell Home Assistant to round the humidity value
When using the value to draw a graph of e.g. the humidity of a crawl
space where you have a dehumidifier installed, the graph becomes very
bumpy if you round the value to whole percentages.  You get a smoother
graph if you allow the value to include the decimal as well, and can
detect trends earlier.