Ensure the announced Home Assistant MQTT device name is constant
When Home Assistant MQTT Discovery is enabled, each ANAVI Thermometer
produces multiple sensors.  All of them are linked to the ANAVI
Thermometer device since they share the same information in the
"device" part of the discovery message.

But there is a bug: the "name" of the device is set to the name of the
sensor.  Luckily Home Assistant still merges all the sensor to the
same device (I believe it only uses the "identifiers" setting to
detect which device the sensor belongs to).  However, depending on
which sensor announcement Home Assistant sees first, the name of the
device may be something like:

    Bedroom Temperature
    Bedroom Humidity
    Bedroom Water Temp
    Bedroom BMP180 Air Pressure

et c.  This commit ensures the name of the device is the same for all
the sensors, so that you end up with a device named like:

    Bedroom ANAVI Thermometer