Make MQTT mode selection code more readable
The USE_MULTIPLE_MQTT and USE_MULTIPLE_STATUS_TOPICS interact with
each other, making the code a bit hard to understand.  I currently
think it is easier to talk about 3 supported modes of operation, and
have 3 separate preprocessor symbols that the code can use to check
which mode is selected:

  MQTT_MODE_MULTIPLE
  MQTT_MODE_MIXED
  MQTT_MODE_SINGLE

Additionally, to make the code more readable, only use "#ifdef" or "if
defined()", avoid using "#ifndef" or "#else" when testing these
mode-selecting macros.

Unless I made a mistake, this commit should not modify the behaviour
in any way.