Add a display language
This commit adds a new display language that can draw lines, dots, and
text to the display.  It can of course also display the measured
values of the various sensors you can connect to the ANAVI
Thermometer.  It also adds 10 user-defined values that can be set via
MQTT, and then displayed.

The display language is based on one-character commands.  Some of them
take one or two numeric prefix arguments.  For example, "C" clears the
display, "1f" selects font number 1, "10V" displays the value of
sensor 10 (which is the DHT22 temperature), and "1,10V" displays the
same temperature but without the trailing "F" or "°C".  DISPLAY.md
documents the available codes.

Remove the legacy code that updated the display, and use a built-in
format string that achieves the same result instead.  Also add two
alternative formats that can be selected via MQTT (and a few test/demo
formats).

The line1, line2 and line3 MQTT messages now set user-defined values
0, 1 and 2.  The new value/N message can set any of the user-defined
values, to either a fixed string or so that they are dynamically
updated from a connected DS18B20 sensor.  The latter only works if
compiled with MULTI_DS18B20_SUPPORT.

The display message selects from the built-in format strings.  The
graph/N messages (for N in 0..9) are combined to form a user-supplied
format string that (if non-empty) overrides the built-in format
string.

Rename setDefaultSensorLines() to printDHT22(), since it now only
prints the air and humidity values to the serial line.