Add limited UTF-8 support to the display language, so we can display "°C"
This fixes a long-standing issue where the temperature would be
reported as e.g. "20.1C" instead of "20.1°C".  To fix this we have to
do two things:

 - Switch fonts from the "_tr" family to "_tf", so that the "°" glyph
   is included.
 - Change Grapher::display_string so that it decodes the UTF-8-encoded
   "°" from formatTemperature() into the ASCII that u8g2.drawStr()
   expects.

This commit also adds a new way to specify a string in the display
language: "[...]" outputs the UTF-8-encoded string "...".  This is
admittedly more convenient to write than the Hollerith Constants, so
update the built-in formats to use it.

Hollerith Constants are still supported, as they are the only way to
display a literal "]" character.