Add support for inverting the display
If the display format contains "1i" in an active part of the format
string, the display will be inverted.  "0i" (or just "i") restores the
normal mode.

This feature uses u8g2.sendF(), which is not available in version
2.23.18 of U8g2.  I have tested this with version 2.34.22, so update
the requirement to that version.  (I have not checked exactly which
version introduced it, so it is possible that some older versions also
support it.)