Add burn-in protection
The new display code "B" prevents display burn-in by automatically
inverting the display half of the time.  How often it switches depends
on if the screen format uses the {...} construct to display alternate
information.  It tries to ensure that all strings are displayed
inverted half the time and in normal mode half the time, to ensure all
pixels are turned on approximately the same amount of time.

This is not used in the default built-in program.  Displays that have
been active since the ANAVI Thermometer were launched are hopelessly
damaged by now, and will be totally unreadable in inverted mode, so it
is better to just live with the damage and use normal mode all the
time.

But if you purchase a new display, I strongly recommend that you use a
format that includes the "B" feature.