Handle ArduinoJson 6.19.0 and newer; recommend 6.21.4
ArduinoJson 6.19.0 uses double instead of float by default.  This
causes serialized floats to sometimes contain trailing digits that are
rounding errors.  For instance, 21.2 might be sent as "21.20000076".
Fix by defining ARDUINOJSON_USE_DOUBLE to 0, which was the default
value in older versions of ArduinoJson.

Change the recommended version to the current release: 6.21.4.